fix(ocpp): route wire-log encoding through the error-logging helpers
Move encode_action_or_log/encode_response_or_log out of client/v2_common
into a shared ocpp/wire_log module so both roles can reach them without
cross-role coupling, then route the ~13 server/client handler and backend
sites that encoded message-log payloads with .unwrap_or(Value::Null) through
them. Encode failures now log to stderr before degrading to null instead of
vanishing silently (OC-R-101).
